+++ CHANGELOG.md
@@ -39,6 +39,8 @@
 - 'EmTimeout' classes split to 'EmTimeout', 'EmTimeoutShort' and 'EmTimeoutLong' 
 - 'EmDuration' classes split to 'EmDuration', 'EmDurationShort' and 'EmDurationLong'. 
 - 'EmDuration' added 'to' method to get duration days, hours, minutes, seconds and milliseconds.
+- Added 'EmTask' class to start and handle separate task.
+- Added 'EmAppTaskInterfaces' class to allow one or more interfaces to run in a separate task.
 - Added new 'EmTag' & 'EmTags' classes to have simple synchronization among same values coming from different sources
 - Added new 'EmAutoPtr' class as a lightweight alternative to std::auto_ptr for environments where the standard library is not available or desired
 - Removed 'SoftwareSerial' as dependency since its not more used.